requirements.txt
amqp==5.1.1
asgiref==3.6.0
base==0.0.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==2.1.0
Django==4.2.1
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-paypal==2.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.12.0
fitz==0.0.1.dev2
food==0.0.11
future==0.18.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
kombu==5.3.1
looseversion==1.3.0
lxml==4.9.3
MarkupPy==1.14
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
paypalrestsdk==1.13.1
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.39
prov==2.0.0
pycparser==2.21
pycryptodome==3.18.0
pydot==1.4.2
PyJWT==2.7.0
PyMuPDF==1.22.5
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.6
PyYAML==6.0.1
razorpay==1.3.0
rdflib==7.0.0
requests==2.31.0
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tablib==3.5.0
traits==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
